Updated the project to include the MPU6050_Kalman and automatic mode
With the MPU6050_Kalman library added to the project,
the main.cpp file was updated to include the library
header file and use the library functions to filter
the raw accelerometer and gyroscope data from the
MPU6050 sensor.

Further, the main.cpp file was updated to the first
version of automatic mode. The automatic mode will
be used to automatically control the mirrors based
on the sensor position and the desired position
received from the mqtt server.